+++ src/etherscan.ts
@@ -127,6 +127,18 @@ export async function submitSources(
     case '42':
       host = 'https://api-kovan.etherscan.io';
       break;
+    case '97':
+      host = 'https://api-testnet.bscscan.com';
+      break;
+    case '56':
+      host = 'https://api.bscscan.com';
+      break;
+    case '128':
+      host = 'https://api.hecoinfo.com';
+      break;
+    case '256':
+      host = 'https://api-testnet.hecoinfo.com';
+      break;
     default:
       return logError(`Network with chainId: ${chainId} not supported`);
   }