+++ src/helpers.ts
@@ -7,11 +7,13 @@ import {
   ContractFactory,
   PayableOverrides,
 } from '@ethersproject/contracts';
+import {AddressZero} from '@ethersproject/constants';
 import {BigNumber} from '@ethersproject/bignumber';
 import {Wallet} from '@ethersproject/wallet';
 import {keccak256 as solidityKeccak256} from '@ethersproject/solidity';
 import {zeroPad, hexlify} from '@ethersproject/bytes';
 import {Interface, FunctionFragment} from '@ethersproject/abi';
+import {LedgerSigner} from '@ethersproject/hardware-wallets';
 import {
   Deployment,
   DeployResult,
@@ -45,6 +47,7 @@ import {
   EthereumProvider,
   HardhatRuntimeEnvironment,
 } from 'hardhat/types';
+import {DeploymentsManager} from './DeploymentsManager';
 
 diamondBase.abi = mergeABIs(
   [
@@ -169,6 +172,7 @@ let provider: Web3Provider;
 const availableAccounts: {[name: string]: boolean} = {};
 export function addHelpers(
   env: HardhatRuntimeEnvironment,
+  deploymentManager: DeploymentsManager,
   partialExtension: PartialExtension, // TODO
   getArtifact: (name: string) => Promise<Artifact>,
   saveDeployment: (
@@ -193,6 +197,10 @@ export function addHelpers(
         const accounts = await provider.send('eth_accounts', []);
         for (const account of accounts) {
           availableAccounts[account.toLowerCase()] = true;
+        }
+
+        for (const address of deploymentManager.impersonatedAccounts) {
+          availableAccounts[address.toLowerCase()] = true;
         }
       } catch (e) {}
     }
@@ -252,7 +260,7 @@ export function addHelpers(
     from: string;
     log?: boolean;
   }): Promise<string> {
-    const {address: from, ethersSigner} = getFrom(options.from);
+    const {address: from, ethersSigner, hardwareWallet} = getFrom(options.from);
     if (!ethersSigner) {
       throw new Error('no signer for ' + from);
     }
@@ -261,27 +269,40 @@ export function addHelpers(
     if (code === '0x') {
       const senderAddress = '0x3fab184622dc19b6109349b94811493bf2a45362';
       // TODO gasPrice override
+
+      if (options.log || hardwareWallet) {
+        print(
+          `sending eth to create2 contract deployer address (${senderAddress})`
+        );
+        if (hardwareWallet) {
+          print(` (please confirm on your ${hardwareWallet})`);
+        }
+      }
       const ethTx = await ethersSigner.sendTransaction({
         to: senderAddress,
         value: BigNumber.from('10000000000000000').toHexString(),
       });
-      if (options.log) {
-        log(
-          `sending eth to create2 contract deployer address (${senderAddress}) (tx: ${ethTx.hash})...`
-        );
+      if (options.log || hardwareWallet) {
+        log(` (tx: ${ethTx.hash})...`);
       }
       await ethTx.wait();
 
       // await provider.send("eth_sendTransaction", [{
       //   from
       // }]);
+      if (options.log || hardwareWallet) {
+        print(
+          `deploying create2 deployer contract (at ${create2DeployerAddress}) using deterministic deployment (https://github.com/Arachnid/deterministic-deployment-proxy)`
+        );
+        if (hardwareWallet) {
+          print(` (please confirm on your ${hardwareWallet})`);
+        }
+      }
       const deployTx = await provider.sendTransaction(
         '0xf8a58085174876e800830186a08080b853604580600e600039806000f350fe7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe03601600081602082378035828234f58015156039578182fd5b8082525050506014600cf31ba02222222222222222222222222222222222222222222222222222222222222222a02222222222222222222222222222222222222222222222222222222222222222'
       );
-      if (options.log) {
-        log(
-          `deploying create2 deployer contract (at ${create2DeployerAddress}) using deterministic deployment (https://github.com/Arachnid/deterministic-deployment-proxy) (tx: ${deployTx.hash})...`
-        );
+      if (options.log || hardwareWallet) {
+        log(` (tx: ${deployTx.hash})...`);
       }
       await deployTx.wait();
     }
@@ -330,7 +351,7 @@ export function addHelpers(
   ): Promise<DeployResult> {
     const args: any[] = options.args ? [...options.args] : [];
     await init();
-    const {address: from, ethersSigner} = getFrom(options.from);
+    const {address: from, ethersSigner, hardwareWallet} = getFrom(options.from);
     if (!ethersSigner) {
       throw new Error('no signer for ' + from);
     }
@@ -387,10 +408,17 @@ export function addHelpers(
       ethersSigner.estimateGas(newOverrides)
     );
     await setupGasPrice(unsignedTx);
+
+    if (options.log || hardwareWallet) {
+      print(`deploying "${name}"`);
+      if (hardwareWallet) {
+        print(` (please confirm on your ${hardwareWallet})`);
+      }
+    }
     let tx = await ethersSigner.sendTransaction(unsignedTx);
 
-    if (options.log) {
-      print(`deploying "${name}" (tx: ${tx.hash})...`);
+    if (options.log || hardwareWallet) {
+      print(` (tx: ${tx.hash})...`);
     }
 
     // await overrideGasLimit(overrides, options, newOverrides =>
@@ -436,7 +464,7 @@ export function addHelpers(
       libraries: options.libraries,
     };
     await saveDeployment(name, deployment);
-    if (options.log) {
+    if (options.log || hardwareWallet) {
       print(
         `: deployed at ${deployment.address} with ${receipt?.gasUsed} gas\n`
       );
@@ -590,14 +618,11 @@ export function addHelpers(
       }
 
       if (transaction) {
+        const {ethersSigner} = await getOptionalFrom(options.from);
         const {artifact} = await getArtifactFromOptions(name, options);
         const abi = artifact.abi;
         const byteCode = linkLibraries(artifact, options.libraries);
-        const factory = new ContractFactory(
-          abi,
-          byteCode,
-          provider.getSigner(options.from)
-        );
+        const factory = new ContractFactory(abi, byteCode, ethersSigner);
 
         const compareOnData = fieldsToCompareArray.indexOf('data') !== -1;
 
@@ -649,7 +674,32 @@ export function addHelpers(
       } else {
         const deployment = await getDeploymentOrNUll(name);
         if (deployment) {
-          result = deployment as DeployResult;
+          if (
+            options.deterministicDeployment &&
+            diffResult.address &&
+            diffResult.address.toLowerCase() !== deployment.address
+          ) {
+            const {
+              artifact: linkedArtifact,
+              artifactName,
+            } = await getLinkedArtifact(name, options);
+
+            // receipt missing
+            const newDeployment = {
+              ...linkedArtifact,
+              address: diffResult.address,
+              linkedData: options.linkedData,
+              libraries: options.libraries,
+              args: argsArray,
+            };
+            await saveDeployment(name, newDeployment, artifactName);
+            result = {
+              ...newDeployment,
+              newlyDeployed: false,
+            };
+          } else {
+            result = deployment as DeployResult;
+          }
         } else {
           if (!diffResult.address) {
             throw new Error(
@@ -870,9 +920,15 @@ Plus they are only used when the contract is meant to be used as standalone when
             `To change owner/admin, you need to call ${changeOwnershipMethod}`
           );
         }
+        if (currentOwner === AddressZero) {
+          throw new Error(
+            'The Proxy belongs to no-one. It cannot be upgraded anymore'
+          );
+        }
+
         const executeReceipt = await execute(
           proxyName,
-          {...options},
+          {...options, from: currentOwner},
           changeImplementationMethod,
           implementation.address,
           data
@@ -887,6 +943,7 @@ Plus they are only used when the contract is meant to be used as standalone when
         address: proxy.address,
         linkedData: options.linkedData,
         abi: mergedABI,
+        implementation: implementation.address,
         args: proxy.args,
         execute: updateMethod
           ? {
@@ -908,6 +965,27 @@ Plus they are only used when the contract is meant to be used as standalone when
         newlyDeployed: true,
       };
     } else {
+      const oldDeployment = await env.deployments.get(name);
+
+      if (oldDeployment.implementation !== implementation.address) {
+        const proxiedDeployment: DeploymentSubmission = {
+          ...oldDeployment,
+          implementation: implementation.address,
+          linkedData: options.linkedData,
+          abi: mergedABI,
+          execute: updateMethod
+            ? {
+                methodName: updateMethod,
+                args: argsArray,
+              }
+            : undefined,
+        };
+        proxiedDeployment.history = proxiedDeployment.history
+          ? proxiedDeployment.history.concat([oldDeployment])
+          : [oldDeployment];
+        await saveDeployment(name, proxiedDeployment);
+      }
+
       const deployment = await env.deployments.get(name);
       return {
         ...deployment,
@@ -926,19 +1004,26 @@ Plus they are only used when the contract is meant to be used as standalone when
 
   function getOptionalFrom(
     from?: string
-  ): {address?: Address; ethersSigner?: Signer} {
+  ): {address?: Address; ethersSigner?: Signer; hardwareWallet?: string} {
     return _getFrom(from, true);
   }
 
-  function getFrom(from?: string): {address: Address; ethersSigner?: Signer} {
-    return _getFrom(from, false) as {address: Address; ethersSigner?: Signer};
+  function getFrom(
+    from?: string
+  ): {address: Address; ethersSigner?: Signer; hardwareWallet?: string} {
+    return _getFrom(from, false) as {
+      address: Address;
+      ethersSigner?: Signer;
+      hardwareWallet?: string;
+    };
   }
 
   function _getFrom(
     from?: string,
     optional?: boolean
-  ): {address?: Address; ethersSigner?: Signer} {
+  ): {address?: Address; ethersSigner?: Signer; hardwareWallet?: string} {
     let ethersSigner: Signer | undefined;
+    let hardwareWallet: string | undefined = undefined;
     if (!from) {
       if (optional) {
         return {};
@@ -955,10 +1040,17 @@ Plus they are only used when the contract is meant to be used as standalone when
     } else {
       if (availableAccounts[from.toLowerCase()]) {
         ethersSigner = provider.getSigner(from);
+      } else {
+        // TODO register protocol based account as availableAccounts ? if so do not else here
+        if (
+          deploymentManager.addressesToProtocol[from.toLowerCase()] === 'ledger'
+        )
+          ethersSigner = new LedgerSigner(provider);
+        hardwareWallet = 'ledger';
       }
     }
 
-    return {address: from, ethersSigner};
+    return {address: from, ethersSigner, hardwareWallet};
   }
 
   // async function findEvents(contract: Contract, event: string, blockHash: string): Promise<any[]> {
@@ -991,7 +1083,7 @@ Plus they are only used when the contract is meant to be used as standalone when
     }
 
     const proxyName = name + '_DiamondProxy';
-    const {address: owner} = getProxyOwner(options);
+    const {address: owner, hardwareWallet} = getProxyOwner(options);
     const facetSnapshot: FacetCut[] = [];
     const oldFacets: FacetCut[] = [];
     if (oldDeployment) {
@@ -1193,7 +1285,7 @@ Plus they are only used when the contract is meant to be used as standalone when
             throw new Error('DiamondCreated Not Emitted');
           }
           const proxyAddress = diamondCreatedEvent.args.diamond;
-          if (options.log) {
+          if (options.log || hardwareWallet) {
             log(
               `Diamond deployed at ${proxyAddress} via Diamantaire (${diamantaireDeployment.address} (tx: ${createReceipt.transactionHash})) with ${createReceipt.gasUsed} gas`
             );
@@ -1301,7 +1393,7 @@ Plus they are only used when the contract is meant to be used as standalone when
   async function rawTx(tx: SimpleTx): Promise<Receipt> {
     tx = {...tx};
     await init();
-    const {address: from, ethersSigner} = getFrom(tx.from);
+    const {address: from, ethersSigner, hardwareWallet} = getFrom(tx.from);
     if (!ethersSigner) {
       throw new UnknownSignerError({
         from,
@@ -1318,6 +1410,9 @@ Plus they are only used when the contract is meant to be used as standalone when
         nonce: tx.nonce,
         data: tx.data,
       };
+      if (hardwareWallet) {
+        log(` please confirm on your ${hardwareWallet}`);
+      }
       let pendingTx = await ethersSigner.sendTransaction(transactionData);
       pendingTx = await onPendingTx(pendingTx);
       if (tx.autoMine) {
@@ -1330,8 +1425,15 @@ Plus they are only used when the contract is meant to be used as standalone when
   }
 
   async function catchUnknownSigner(
-    action: Promise<any> | (() => Promise<any>)
-  ): Promise<void> {
+    action: Promise<any> | (() => Promise<any>),
+    options?: {log?: boolean}
+  ): Promise<null | {
+    from: string;
+    to?: string;
+    value?: string;
+    data?: string;
+  }> {
+    const outputLog = !options || options.log === undefined || options.log;
     try {
       if (action instanceof Promise) {
         await action;
@@ -1341,42 +1443,59 @@ Plus they are only used when the contract is meant to be used as standalone when
     } catch (e) {
       if (e instanceof UnknownSignerError) {
         const {from, to, data, value, contract} = e.data;
-        console.error('no signer for ' + from);
-        if (contract) {
-          console.log(`Please execute the following on ${contract.name}`);
+        if (outputLog) {
           console.log(
-            `
-from: ${from}
-to: ${to}
-value: ${value ? (typeof value === 'string' ? value : value.toString()) : '0'}
-data: ${data}
-`
+            `---------------------------------------------------------------------------------------`
           );
-          console.log('if you have an interface use the following');
+          console.error('no signer for ' + from);
+          console.log(`Please execute the following:`);
           console.log(
-            `
+            `---------------------------------------------------------------------------------------`
+          );
+          if (contract) {
+            console.log(
+              `
 from: ${from}
-to: ${to} (${contract.name})
-value: ${value}
+to: ${to} (${contract.name})${
+                value
+                  ? '\nvalue: ' +
+                    (typeof value === 'string' ? value : value.toString())
+                  : ''
+              }
 method: ${contract.method}
-args: [${contract.args.join(',')}]
+args:
+  - ${contract.args.join('\n  - ')}
+
+(raw data: ${data} )
 `
-          );
-        } else {
-          console.log(`Please execute the following`);
-          console.log(
-            `
+            );
+          } else {
+            console.log(
+              `
 from: ${from}
-to: ${to}
-value: ${value ? (typeof value === 'string' ? value : value.toString()) : '0'}
+to: ${to}${
+                value
+                  ? '\nvalue: ' +
+                    (typeof value === 'string' ? value : value.toString())
+                  : ''
+              }
 data: ${data}
 `
+            );
+          }
+          console.log(
+            `---------------------------------------------------------------------------------------`
           );
         }
+        if (!value || typeof value === 'string') {
+          return {from, to, value, data};
+        }
+        return {from, to, value: value?.toString(), data};
       } else {
         throw e;
       }
     }
+    return null;
   }
 
   async function execute(
@@ -1387,7 +1506,7 @@ data: ${data}
   ): Promise<Receipt> {
     options = {...options}; // ensure no change
     await init();
-    const {address: from, ethersSigner} = getFrom(options.from);
+    const {address: from, ethersSigner, hardwareWallet} = getFrom(options.from);
 
     let tx;
     const deployment = await env.deployments.get(name);
@@ -1416,6 +1535,13 @@ data: ${data}
       throw new Error(
         `expected ${numArguments} arguments for method "${methodName}", got ${args.length}`
       );
+    }
+
+    if (options.log || hardwareWallet) {
+      print(`executing ${name}.${methodName}`);
+      if (hardwareWallet) {
+        print(` (please confirm on your ${hardwareWallet})`);
+      }
     }
 
     if (!ethersSigner) {
@@ -1449,8 +1575,8 @@ data: ${data}
     }
     tx = await onPendingTx(tx);
 
-    if (options.log) {
-      print(`executing ${name}.${methodName} (tx: ${tx.hash}) ...`);
+    if (options.log || hardwareWallet) {
+      print(` (tx: ${tx.hash}) ...`);
     }
 
     if (options.autoMine) {
@@ -1459,7 +1585,7 @@ data: ${data}
       } catch (e) {}
     }
     const receipt = await tx.wait();
-    if (options.log) {
+    if (options.log || hardwareWallet) {
       print(`: performed with ${receipt.gasUsed} gas\n`);
     }
     return receipt;